+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+
+- `Client::with_config` allowing to create a `Resend` client with custom configuration
+  (https://github.com/resend/resend-rust/pull/28)
 
 ## [0.15.0] - 2025-05-09